Moving type, fn to useRecentActivity and use undefined
