+++ src/custom/components/Popups/FollowPendingTxPopup/index.tsx
@@ -1,8 +1,7 @@
 import React, { useEffect, useMemo, useCallback, useRef } from 'react'
 
 import { useIsExpertMode } from '@src/state/user/hooks'
-import useRecentActivity, { TransactionAndOrder } from 'hooks/useRecentActivity'
-import { Order } from 'state/orders/actions'
+import { useRecentActivityLastPendingOrder, AddedOrder } from 'hooks/useRecentActivity'
 import {
   useUpdateAtom,
   useAtomValue,
@@ -14,29 +13,12 @@ import {
   handleCloseOrderPopupAtom,
 } from 'state/application/atoms'
 import FollowPendingTxPopupUI from './FollowPendingTxPopupUI'
-import { isPending } from 'components/Web3Status'
-import { OrderID } from '@src/custom/api/gnosisProtocol'
-
-interface AddedOrder extends Order {
-  addedTime: number
-}
-
-function isAnOrder(element: TransactionAndOrder): element is AddedOrder {
-  return 'inputToken' in element && 'outputToken' in element
-}
+import { OrderID } from 'api/gnosisProtocol'
 
 export const useLastPendingOrder = (): { lastPendingOrder: AddedOrder | null; onClose: () => void } => {
   const setShowFollowPendingTxPopup = useUpdateAtom(handleFollowPendingTxPopupAtom)
   const setLastOrderClosed = useUpdateAtom(handleCloseOrderPopupAtom)
-  const allRecentActivity = useRecentActivity()
-  const lastPendingOrder = useMemo(() => {
-    const pendings = allRecentActivity.filter(isPending)
-    const lastOrder = pendings[pendings.length - 1]
-
-    if (!lastOrder || !isAnOrder(lastOrder)) return null
-
-    return lastOrder
-  }, [allRecentActivity])
+  const lastPendingOrder = useRecentActivityLastPendingOrder()
 
   const onClose = useCallback(() => {
     lastPendingOrder && setLastOrderClosed(lastPendingOrder.id)
@@ -63,19 +45,19 @@ export function useCloseFollowTxPopupIfNotPendingOrder() {
 
 const useShowingPopupFirstTime = (orderId: OrderID | undefined) => {
   const showingPopup = useAtomValue(showFollowTxPopupAtom)
-  const _orderPopupHasBeenClosed = useMemo(
+  const _firstTimePopupOrderAppears = useMemo(
     () =>
       selectAtom(followPendingTxPopupAtom, ({ lastOrderPopupClosed }) => {
-        if (!orderId) return
+        if (!orderId) return false
 
-        return lastOrderPopupClosed === orderId
+        return lastOrderPopupClosed !== orderId
       }),
     [orderId]
   )
 
-  const orderPopupHasBeenClosed = useAtomValue(_orderPopupHasBeenClosed)
+  const firstTimePopupOrderAppears = useAtomValue(_firstTimePopupOrderAppears)
 
-  return { showPopup: !orderPopupHasBeenClosed && showingPopup, orderPopupHasBeenClosed }
+  return { showPopup: firstTimePopupOrderAppears && showingPopup, firstTimePopupOrderAppears }
 }
 
 const FollowPendingTxPopup: React.FC = ({ children }): JSX.Element => {
@@ -83,15 +65,15 @@ const FollowPendingTxPopup: React.FC = ({ children }): JSX.Element => {
   const setHidePendingTxPopupPermanently = useUpdateAtom(handleHidePopupPermanentlyAtom)
   const isExpertMode = useIsExpertMode()
   const { lastPendingOrder, onClose } = useLastPendingOrder()
-  const { showPopup: showFollowPendingTxPopup, orderPopupHasBeenClosed } = useShowingPopupFirstTime(
+  const { showPopup: showFollowPendingTxPopup, firstTimePopupOrderAppears } = useShowingPopupFirstTime(
     lastPendingOrder?.id
   )
 
   useEffect(() => {
-    if (isExpertMode && lastPendingOrder && !orderPopupHasBeenClosed) {
+    if (isExpertMode && lastPendingOrder && firstTimePopupOrderAppears) {
       setShowFollowPendingTxPopup(true)
     }
-  }, [isExpertMode, lastPendingOrder, orderPopupHasBeenClosed, setShowFollowPendingTxPopup])
+  }, [isExpertMode, lastPendingOrder, firstTimePopupOrderAppears, setShowFollowPendingTxPopup])
 
   return (
     <FollowPendingTxPopupUI
